Fix noninteractive optional requires
The assumption here is that in an interactive session, viewing the file
would be enough to load these packages if they're used for code as long
as native Org fontification is set, or they're otherwise autoloaded.

In a noninteractive session on the other hand, it may be worth trying to
load them explicitly.